Projeto atualizado com lancamentos

+++ app/pages/contas/contas.js
@@ -1,4 +1,4 @@
-import {Page, Modal, NavController} from 'ionic-angular';
+import {Page, Modal, NavController, Alert} from 'ionic-angular';
 import {DAOContas} from '../../dao/dao-contas';
 import {ModalContasPage} from '../modal-contas/modal-contas'
 import {Toast} from 'ionic-native';
@@ -13,11 +13,11 @@ export class ContasPage {
 	}
 
 	constructor(nav){
+		this.nav = nav;
 		this.dao = new DAOContas();
 		this.dao.getList((list) =>{
 			this.listContas = list;
 		});
-		this.nav = nav;
 	}
 
 	insert(){
@@ -27,6 +27,10 @@ export class ContasPage {
 			if(data){
 				this.dao.insert(data, (list) =>{
 					this.listContas.push(list);
+
+					Toast.showShortBottom("Conta inserida com sucesso.").subscribe((toast) =>{
+						console.log(toast);
+					});
 				});
 			}
 		});
@@ -39,7 +43,11 @@ export class ContasPage {
 
 		modal.onDismiss((data) =>{
 			if(data){
-				this.dao.edit(data);
+				this.dao.edit(data, (list) =>{
+					Toast.showShortBottom("Conta alterada com sucesso.").subscribe((toast) =>{
+						console.log(toast);
+					});
+				});
 			}
 		});
 
@@ -47,9 +55,27 @@ export class ContasPage {
 	}
 
 	remove(conta){
-		this.dao.remove(conta, (list) =>{
-			let pos = this.listContas.indexOf(list);
-			this.listContas.splice(pos, 1);
+		let confirm = Alert.create({
+			title: "Excluir",
+			body: "Deseja realmente excluir a conta "+conta.description+"?",
+			buttons: [
+				{
+					text: "Sim",
+					handler: () =>{
+						this.dao.remove(conta, (list) =>{
+							let pos = this.listContas.indexOf(list);
+							this.listContas.splice(pos, 1);
+
+							Toast.showShortBottom("Conta removida com sucesso.").subscribe((toast) =>{
+								console.log(toast);
+							});
+						});
+					}
+				},
+				{text: "Não"}
+			]
 		});
+
+		this.nav.present(confirm);
 	}
 }
